+++ roleplay/vexutils/__init__.py
@@ -1,0 +1,8 @@
+from typing import TYPE_CHECKING, Optional
+
+import discord
+from redbot.core.bot import Red
+
+from .chat import humanize_bytes, inline_hum_list, no_colour_rich_markup
+from .meta import format_help, format_info, get_vex_logger, out_of_date_check
+from .version import __version__